Refactor reaction handling and introduce pagination
- Added `PaginationResponseWrapper<T>` for paginated responses.
- Introduced `PostReactionEventDTO` for post reaction events.
- Removed `ReactionPublishDTO` and consolidated `ReactionPost` into `PostReaction`.
- Updated repository and service interfaces to use new classes and response types.
- Refactored controllers to handle new response structures and methods.
- Enhanced dependency injection in `Program.cs` for new services.

+++ react-service/react-service/react-service.Application/Services/ReactionPostService.cs
@@ -1,48 +1,40 @@
 using AutoMapper;
 using Microsoft.Extensions.Options;
-using Microsoft.VisualBasic;
-using react_service.Domain.Entites;
-using react_service.Domain.Enums;
-using react_service.Application.DTO.RabbitMQ;
+using react_service.Application.DTO;
 using react_service.Application.DTO.ReactionPost.Request;
-using react_service.Application.DTO.ReactionPost.Response;
 using react_service.Application.Helpers;
 using react_service.Application.Interfaces.Publishers;
 using react_service.Application.Interfaces.Repositories;
 using react_service.Application.Interfaces.Services;
 using react_service.Application.Pagination;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using static System.Collections.Specialized.BitVector32;
-using react_service.Application.DTO;
+using react_service.Domain.Entites;
 
 namespace react_service.Application.Services
 {
     public class ReactionPostService : IReactionPostService
     {
         // private readonly HttpClient _httpClient; // Remove or comment out this unused field
-        private readonly IReactionPostRepository reactionRepository;
+        private readonly IPostReactionRepository reactionRepository;
+        private readonly IPostRepository postRepository;
         private readonly IMapper mapper;
         private readonly IOptions<PaginationSettings> paginationSetting;
         private readonly IReactionPublisher reactionPublisher;
 
-        public ReactionPostService(IReactionPostRepository reactionRepository, IMapper mapper, IOptions<PaginationSettings> paginationSetting
+        public ReactionPostService(IPostReactionRepository reactionRepository, IPostRepository postRepository, IMapper mapper, IOptions<PaginationSettings> paginationSetting
             , IReactionPublisher reactionPublisher)
         {
             this.reactionRepository = reactionRepository;
+            this.postRepository = postRepository;
             this.mapper = mapper;
             this.paginationSetting = paginationSetting;
             this.reactionPublisher = reactionPublisher;
         }
 
-        
 
-        public async Task<ResponseWrapper<object>> DeleteReactionAsync(string postId, string userId)
+
+        public async Task<ResponseWrapper<bool>> DeleteReactionAsync(string postId, string userId)
         {
-            var response = new ResponseWrapper<object>();
+            var response = new ResponseWrapper<bool>();
             if (string.IsNullOrEmpty(postId))
             {
                 response.Errors.Add("Post ID cannot be null or empty.");
@@ -55,6 +47,13 @@ namespace react_service.Application.Services
                 response.ErrorType = ErrorType.BadRequest;
                 return response;
             }
+            var postDeleted = await postRepository.IsPostDeleted(postId);
+            if (postDeleted)
+            {
+                response.Errors.Add("Post deleted or doesn't exist");
+                response.ErrorType = ErrorType.BadRequest;
+                return response;
+            }
             var deleted = await reactionRepository.DeleteReactionAsync(postId, userId);
             if (!deleted)
             {
@@ -63,12 +62,13 @@ namespace react_service.Application.Services
                 return response;
             }
             response.Message = "Reaction deleted successfully.";
+            response.Data = true;
             return response;
         }
 
-        public async Task<ResponseWrapper<object>> AddReactionAsync(CreateReactionRequest reaction, string userId)
+        public async Task<ResponseWrapper<bool>> AddReactionAsync(CreateReactionRequest reaction, string userId)
         {
-            var response = new ResponseWrapper<object>();
+            var response = new ResponseWrapper<bool>();
             if (string.IsNullOrEmpty(reaction.PostId))
             {
                 response.Errors.Add("Post ID cannot be null or empty.");
@@ -81,19 +81,34 @@ namespace react_service.Application.Services
                 response.ErrorType = ErrorType.BadRequest;
                 return response;
             }
-            var reactionObj = mapper.Map<ReactionPost>(reaction);
+            var postDeleted = await postRepository.IsPostDeleted(reaction.PostId);
+            if (postDeleted)
+            {
+                response.Errors.Add("Post deleted or doesn't exist");
+                response.ErrorType = ErrorType.BadRequest;
+                return response;
+            }
+            var reactionObj = mapper.Map<PostReaction>(reaction);
             reactionObj.UserId = userId;
             await reactionRepository.AddReactionAsync(reactionObj);
             response.Message = "Reaction added successfully.";
+            response.Data = true;
             return response;
         }
 
-        public async Task<ResponseWrapper<object>> DeleteReactionsByPostId(string postId)
+        public async Task<ResponseWrapper<bool>> DeleteReactionsByPostId(string postId)
         {
-            var response = new ResponseWrapper<object>();
+            var response = new ResponseWrapper<bool>();
             if (string.IsNullOrEmpty(postId))
             {
                 response.Errors.Add("Post ID cannot be null or empty.");
+                response.ErrorType = ErrorType.BadRequest;
+                return response;
+            }
+            var postDeleted = await postRepository.IsPostDeleted(postId);
+            if (postDeleted)
+            {
+                response.Errors.Add("Post deleted or doesn't exist");
                 response.ErrorType = ErrorType.BadRequest;
                 return response;
             }
@@ -108,9 +123,9 @@ namespace react_service.Application.Services
             return response;
         }
 
-        public async Task<ResponseWrapper<PostsReactedByUserDTO>> FilterPostsReactedByUserAsync(List<string> postIds, string userId)
+        public async Task<ResponseWrapper<List<string>>> FilterPostsReactedByUserAsync(List<string> postIds, string userId)
         {
-            var response = new ResponseWrapper<PostsReactedByUserDTO>();
+            var response = new ResponseWrapper<List<string>>();
             if (postIds == null || postIds.Count == 0)
             {
                 response.Errors.Add("Post IDs cannot be null or empty.");
@@ -123,15 +138,15 @@ namespace react_service.Application.Services
                 response.ErrorType = ErrorType.UnAuthorized;
                 return response;
             }
-            var obj = await reactionRepository.FilterPostsReactedByUserAsync(postIds, userId);
-            response.Data = new PostsReactedByUserDTO { postIds = obj };
+            var Ids = await reactionRepository.FilterPostsReactedByUserAsync(postIds, userId);
+            response.Data = Ids ?? new List<string>();
             response.Message = "Filtered posts reacted by user successfully.";
             return response;
         }
 
-        public async Task<ResponseWrapper<PagedReactsResponse>> GetPostsReactedByUserAsync(string userId, string nextReactIdHash)
+        public async Task<PaginationResponseWrapper<List<string>>> GetPostsReactedByUserAsync(string userId, string nextReactIdHash)
         {
-            var response = new ResponseWrapper<PagedReactsResponse>();
+            var response = new PaginationResponseWrapper<List<string>>();
             if (string.IsNullOrEmpty(userId))
             {
                 response.Errors.Add("User ID cannot be null or empty.");
@@ -141,33 +156,40 @@ namespace react_service.Application.Services
             string lastSeenId = string.IsNullOrWhiteSpace(nextReactIdHash) ? "" : PaginationHelper.DecodeCursor(nextReactIdHash!);
             var reactionList = (await reactionRepository.GetPostsReactedByUserAsync(userId, lastSeenId)).ToList();
             bool hasMore = reactionList.Count > (paginationSetting.Value.DefaultPageSize - 1);
-            var pagedResponse = mapper.Map<PagedReactsResponse>(reactionList);
             var lastId = hasMore ? reactionList.Last().Id : null;
-            pagedResponse.HasMore = hasMore;
-            pagedResponse.Next = lastId != null ? PaginationHelper.GenerateCursor(lastId) : null;
-            response.Data = pagedResponse;
+            response.Data = reactionList.Select(r => r.Id).ToList();
+            response.HasMore = hasMore;
+            response.Next = lastId != null ? PaginationHelper.GenerateCursor(lastId) : null;
+
             response.Message = "Posts reacted by user retrieved successfully.";
             return response;
         }
 
-        public async Task<ResponseWrapper<ReactionsUsersResponse>> GetUserIdsReactedToPostAsync(string postId)
+        public async Task<ResponseWrapper<List<string>>> GetUserIdsReactedToPostAsync(string postId)
         {
-            var response = new ResponseWrapper<ReactionsUsersResponse>();
+            var response = new ResponseWrapper<List<string>>();
             if (string.IsNullOrEmpty(postId))
             {
                 response.Errors.Add("Post ID cannot be null or empty.");
                 response.ErrorType = ErrorType.BadRequest;
                 return response;
             }
+            var postDeleted = await postRepository.IsPostDeleted(postId);
+            if (postDeleted)
+            {
+                response.Errors.Add("Post deleted or doesn't exist");
+                response.ErrorType = ErrorType.BadRequest;
+                return response;
+            }
             var userIds = await reactionRepository.GetUserIdsReactedToPostAsync(postId);
-            response.Data = new ReactionsUsersResponse { UserIds = userIds };
+            response.Data = userIds;
             response.Message = "User IDs retrieved successfully.";
             return response;
         }
 
-        public async Task<ResponseWrapper<ReactionsUsersResponse>> GetUserIdsReactedToPostAsync(string postId, string next, int take)
+        public async Task<PaginationResponseWrapper<List<string>>> GetUserIdsReactedToPostAsync(string postId, string next, int take)
         {
-            var response = new ResponseWrapper<ReactionsUsersResponse>();
+            var response = new PaginationResponseWrapper<List<string>>();
             if (string.IsNullOrEmpty(postId))
             {
                 response.Errors.Add("Post ID cannot be null or empty.");
@@ -175,13 +197,60 @@ namespace react_service.Application.Services
                 return response;
             }
             string lastSeenId = string.IsNullOrWhiteSpace(next) ? "" : PaginationHelper.DecodeCursor(next);
+            var postDeleted = await postRepository.IsPostDeleted(postId);
+            if (postDeleted)
+            {
+                response.Errors.Add("Post deleted or doesn't exist");
+                response.ErrorType = ErrorType.BadRequest;
+                return response;
+            }
             var userIds = await reactionRepository.GetUserIdsReactedToPostAsync(postId, lastSeenId, take + 1);
             bool hasMore = userIds.Count > take;
             string nextCursor = hasMore ? PaginationHelper.GenerateCursor(userIds[take]) : null;
             if (hasMore) userIds = userIds.Take(take).ToList();
-            response.Data = new ReactionsUsersResponse { UserIds = userIds, HasMore = hasMore, Next = nextCursor };
+            response.Data = userIds;
+            response.HasMore = hasMore;
+            response.Next = nextCursor;
             response.Message = "User IDs retrieved successfully.";
             return response;
         }
+
+        public async Task<ResponseWrapper<bool>> IsPostReactedByUserAsync(string postId, string userId)
+        {
+            var response = new ResponseWrapper<bool>();
+            if (string.IsNullOrEmpty(postId))
+            {
+                response.Errors.Add("Post ID cannot be null or empty.");
+                response.ErrorType = ErrorType.BadRequest;
+                return response;
+            }
+            if (string.IsNullOrEmpty(userId))
+            {
+                response.Errors.Add("User ID cannot be null or empty.");
+                response.ErrorType = ErrorType.UnAuthorized;
+                return response;
+            }
+            try
+            {
+                var postDeleted = await postRepository.IsPostDeleted(postId);
+                if (postDeleted)
+                {
+                    response.Errors.Add("Post deleted or doesn't exist");
+                    response.ErrorType = ErrorType.BadRequest;
+                    return response;
+                }
+                var isReacted = await reactionRepository.IsPostReactedByUserAsync(postId, userId);
+                response.Data = isReacted;
+                response.Message = "Checked if post is reacted by user successfully.";
+            }
+            catch (Exception ex)
+            {
+                response.Errors.Add($"An error occurred: {ex.Message}");
+                response.ErrorType = ErrorType.InternalServerError;
+                return response;
+            }
+            return response;
+        }
+
     }
 }
